LayoutTests/http/tests/resourceLoadStatistics/partitioned-and-unpartitioned-cookie-with-partitioning-timeout-expected.txt

CONSOLE MESSAGE: line 114: Bootstrapping test.
CONSOLE MESSAGE: line 65: Step 1.
CONSOLE MESSAGE: line 70: Step 2.
CONSOLE MESSAGE: line 77: Step 3.
CONSOLE MESSAGE: line 84: Step 4.
CONSOLE MESSAGE: line 54: setNormalCookiePartitioningTimeOutPlusUserInteractionAndContinue().
CONSOLE MESSAGE: line 91: Step 5.
CONSOLE MESSAGE: line 48: setShortCookiePartitioningTimeOutPlusFireShouldPartitionCookiesHandlerAndContinue().
CONSOLE MESSAGE: line 43: fireShouldPartitionCookiesHandlerAndContinue().
CONSOLE MESSAGE: line 98: Step 6.
CONSOLE MESSAGE: line 54: setNormalCookiePartitioningTimeOutPlusUserInteractionAndContinue().
CONSOLE MESSAGE: line 105: Step 7.
PASS successfullyParsed is true

TEST COMPLETE
  

--------
Frame: '<!--frame1-->'
--------
Should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame2-->'
--------
Setting partitioned, third-party cookie.


--------
Frame: '<!--frame3-->'
--------
Should only receive partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--frame4-->'
--------
After first user interaction, should still only receive a partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--frame5-->'
--------
Should only receive partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--frame6-->'
--------
After second user interaction, should still only receive a partitioned, third-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value
